Renamed files according to standard
