+++ src/CMakeLists.txt
@@ -52,7 +52,7 @@ set(SOURCE_FILES
   ${CMAKE_CURRENT_SOURCE_DIR}/logger.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/model.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/namedentity.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/operatorLibrary.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/operatorlibrary.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/orderedentity.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/parser.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/printer.cpp
@@ -80,7 +80,7 @@ set(API_HEADER_FILES
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/logger.h
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/model.h
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/namedentity.h
-  ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/operatorLibrary.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/operatorlibrary.h
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/orderedentity.h
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/parser.h
   ${CMAKE_CURRENT_SOURCE_DIR}/api/libcellml/printer.h
