+++ src/api/libcellml/operatorlibrary.h
